requirements.txt
asgiref==3.8.1
attrs==25.1.0
Django==5.1.6
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.9
factory_boy==3.3.3
Faker==36.1.1
freezegun==1.5.1
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pyotp==2.9.0
pytest==8.3.5
pytest-django==4.10.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
qrcode==8.0
referencing==0.36.2
rpds-py==0.23.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
Werkzeug==3.1.3
